refactor(MCP-483): wire PrivateLink through connection config registry + fix update test
