+++ tests/unit/tools/atlas/streams/connectionConfigs.test.ts
@@ -7,7 +7,6 @@ import {
     LambdaConnectionConfig,
     HttpsConnectionConfig,
     SchemaRegistryConnectionConfig,
-    PrivateLinkConnectionConfig,
     getConnectionConfigSchema,
 } from "../../../../../src/tools/atlas/streams/connectionConfigs.js";
 
@@ -158,25 +157,6 @@ describe("per-type connection config schemas", () => {
                 schemaRegistryUrls: ["https://sr.example.com"],
                 bootstrapServers: "broker1:9092",
             });
-            expect(result.success).toBe(false);
-        });
-    });
-
-    describe("PrivateLinkConnectionConfig", () => {
-        it("accepts AWS PrivateLink with required fields", () => {
-            const result = PrivateLinkConnectionConfig.safeParse({
-                provider: "AWS",
-                region: "us-east-1",
-                vendor: "CONFLUENT",
-                serviceEndpointId: "com.amazonaws.vpce.us-east-1.vpce-svc-xyz",
-                dnsDomain: "example.confluent.cloud",
-                dnsSubDomain: [],
-            });
-            expect(result.success).toBe(true);
-        });
-
-        it("rejects PrivateLink config without provider", () => {
-            const result = PrivateLinkConnectionConfig.safeParse({ region: "us-east-1" });
             expect(result.success).toBe(false);
         });
     });
